Remove redundant post frontmatter
According to the Zola documentation, the post date and slug can be
automatically inferred when its file/directory name is formatted
`YYYY-mm-dd-title`. This behavior can be disabled by setting
`slugify.paths_keep_dates = true`, but we don't do this. As such, we can
safely remove quite a bit of redundant frontmatter.