Work around missing IPA chars in Fira Code with Noto Sans fallback
This hack is awful, but I think I've found a suitable workaround to Fira
Code lacking glyphs for the IPA (International Phonetic Alphabet)
characters ʔ and ˈ. Testing with Chrome on Android indicates that this
commit fixes the text rendering issues seen in the "about" page.

The "solution" here is to generate a tiny subset of a different
monospace font that supports these glyphs (Noto Sans Mono, in this case)
and rely on CSS's per-character fallback behavior so the text renders
properly. In short, make the browser render everything in Fira Code that
it can _except_ for these two glyphs which render in Noto Sans Mono
instead.

See this upstream issue for details:

https://github.com/tonsky/FiraCode/issues/201